2022 base trajectory params (#83)
* Add 2022 path gen config

Also, fix a few bugs in base trajectory param loading

* Comment out field markers for now

+++ zebROS_ws/src/base_trajectory/include/base_trajectory/obstacle_gradient.h
@@ -28,8 +28,8 @@ class ObstacleGradient
 		void calculatePotentials(const unsigned char *costs, unsigned int *potential);
 		void setSize(int nx, int ny);
 
-		void setThreshold(uint8_t threshold);
-		uint8_t getThreshold(void) const;
+		void setThreshold(int threshold);
+		int getThreshold(void) const;
 	private:
 		std::unique_ptr<ObstacleGradientImpl> impl_;
 };
